python/requirements.txt
requests
bs4
latexcodec
pillow